+++ ContentPatcher/Framework/Tokens/IHigherLevelToken.cs
@@ -1,14 +1,12 @@
 namespace ContentPatcher.Framework.Tokens
 {
     /// <summary>A token wrapped by Content Patcher to handle global input arguments.</summary>
-    /// <typeparam name="TToken">The wrapped token type.</typeparam>
-    internal interface IHigherLevelToken<out TToken> : IToken
-        where TToken : IToken
+    internal interface IHigherLevelToken : IToken
     {
         /*********
         ** Accessors
         *********/
         /// <summary>The wrapped token instance.</summary>
-        TToken Token { get; }
+        IToken Token { get; }
     }
 }