+++ ContentPatcher/release-notes.md
@@ -4,6 +4,7 @@
 ## Upcoming release
 * Added [`patch reload`](docs/author-guide.md#patch-reload) console command (thanks to spacechase0!).
 * Removed legacy token API obsolete since Content Patcher 1.12.
+* Internal changes to prepare for realtime content updates.
 
 ## 1.16.4
 Released 12 August 2020.
@@ -38,7 +39,7 @@ Released 02 August 2020. See the [release highlights for mod authors](https://ww
 * Fixed auto-generated patch names not normalising path separators.
 * Fixed `patch summary` showing duplicate target paths in some cases.
 * Fixed string sorting/comparison for some special characters.
-* Internal changes to prepare for upcoming features.
+* Internal changes to prepare for realtime content updates.
 
 **Update note for mod authors:**  
 Using `"Action": "EditData"` with a `FromFile` field is now deprecated, though it still works.